test: garante que é retornado 403 se não tiver um token de acesso
